+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/core/teleop/ArcadeDrive.java
@@ -1,9 +1,8 @@
-package org.firstinspires.ftc.teamcode.core.linear;
+package org.firstinspires.ftc.teamcode.core.teleop;
 
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.hardware.Servo;
 import com.qualcomm.robotcore.util.Range;
 
 /**
